VQVAe

+++ train_class.py
@@ -75,7 +75,7 @@ def train(epoch, loader, model, optimizer, scheduler, device, entropy_coder, pop
         logits = entropy_coder(id_t.detach().float())
 
         s = s.argmax(-1)
-        s = s.unsqueeze(1)
+        s = s.unsqueeze(1).long()
 
         prior_loss = criterion(logits, s)
